[ViewRepository] Rename query to view repository
