+++ src/Controller/Product/ShowProductDetailsBySlugAction.php
@@ -4,21 +4,21 @@ namespace Sylius\ShopApiPlugin\Controller\Product;
 
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\View\ViewHandlerInterface;
-use Sylius\ShopApiPlugin\Query\ProductDetailsQueryInterface;
+use Sylius\ShopApiPlugin\ViewRepository\ProductDetailsViewRepositoryInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 final class ShowProductDetailsBySlugAction
 {
-    /** @var ProductDetailsQueryInterface */
+    /** @var ProductDetailsViewRepositoryInterface */
     private $productCatalog;
 
     /** @var ViewHandlerInterface */
     private $viewHandler;
 
     public function __construct(
-        ProductDetailsQueryInterface $productCatalog,
+        ProductDetailsViewRepositoryInterface $productCatalog,
         ViewHandlerInterface $viewHandler
     ) {
         $this->productCatalog = $productCatalog;
