+++ src/Resources/config/services/queries.xml
@@ -7,13 +7,13 @@
             <argument type="service" id="sylius.shop_api_plugin.factory.cart_view_factory" />
         </service>
 
-        <service id="sylius.shop_api_plugin.query.product_details_query" class="Sylius\ShopApiPlugin\Query\ProductDetailsQuery">
+        <service id="sylius.shop_api_plugin.view_repository.product_details_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\ProductDetailsViewRepository">
             <argument type="service" id="sylius.repository.channel" />
             <argument type="service" id="sylius.repository.product" />
             <argument type="service" id="sylius.shop_api_plugin.factory.detailed_product_view_factory" />
         </service>
 
-        <service id="sylius.shop_api_plugin.query.product_catalog_query" class="Sylius\ShopApiPlugin\Query\ProductCatalogQuery">
+        <service id="sylius.shop_api_plugin.view_repository.product_catalog_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\ProductCatalogViewRepository">
             <argument type="service" id="sylius.repository.channel" />
             <argument type="service" id="sylius.repository.product" />
             <argument type="service" id="sylius.repository.taxon" />
